feat: 설정 버튼 뷰 제작
