+++ frontend-ios/YakAl/YakAl/Sources/Views/Mypage/MypageSwiftUIView.swift
@@ -144,38 +144,6 @@ struct MypageSwiftUIView: View {
         }
 }
 
-//MARK: - SettingButton View
-struct SettingButton: View {
-    let action: () -> Void
-    let imageName: String?
-    let title: String
-
-    init(action: @escaping () -> Void, imageName: String?, title: String) {
-        self.action = action
-        self.imageName = imageName
-        self.title = title
-    }
-
-    var body: some View {
-        Button(action: action) {
-            HStack {
-                if let validImageName = imageName {  // 옵셔널 바인딩을 사용하여 이미지 이름이 nil이 아닌 경우에만 Image를 표시
-                               Image(validImageName)
-                           }
-                Text(title)
-                    .font(Font.custom("SUIT", size: 16).weight(.medium))
-                    .foregroundColor(Color(red: 0.08, green: 0.08, blue: 0.08))
-                Spacer()
-            }
-            .frame(maxWidth: .infinity)
-            .padding(.horizontal, 20)
-            .padding(.vertical,15)
-            .background(Color(red: 1, green: 1, blue: 1))
-        }
-    }
-}
-
-
 struct MypageSwiftUIView_Previews: PreviewProvider {
     static var previews: some View {
         MypageSwiftUIView()
